\docType{data}
\name{exdata}
\alias{exdata}
\title{A hypothetical simulated cohort dataset based on the GUSTO-1 clinical trial}
\format{A data frame with 2880 rows and 21 variables}
\usage{
data(exdata)
}
\description{
\itemize{
  \item \code{ID}: Identification numbers of individual participants
  \item \code{Y}: An indicator for the occurance of event (Yes=1, No=0)
  \item \code{AGE}: Age
  \item \code{HEI}: Height
  \item \code{WEI}: Weight
  \item \code{SEX}: Gender (Female)
  \item \code{DIA}: Diabetes
  \item \code{HYP}: Hypotension, abnormally low blood pressure (BP < 100)
  \item \code{HRT}: Tachycardia (pulse > 80)
  \item \code{HIG}: High risk
  \item \code{SHO}: Shock (Killip class III/IV)
  \item \code{TTR}: Time to relief > 1 hour
  \item \code{PMI}: Previous MI
  \item \code{HTN}: Hypertension history
  \item \code{LIP}: Hypercholesterolemia
  \item \code{PAN}: Previous angina pectoris
  \item \code{FAM}: Family history of MI
  \item \code{ST4}: ST elevation in >4 leads
  \item \code{A65}: Age > 65 years
  \item \code{SMK1}: Ex-smoker
  \item \code{SMK2}: Corrent smoker
}
}
\references{
Noma, H., Shinozaki, T., Iba, K., Teramukai, S. and Furukawa, T. A. (2020). 
Confidence intervals of prediction accuracy measures for multivariable prediction models based on the bootstrap-based optimism correction methods. 
arXiv:2005.01457.
\url{https://arxiv.org/abs/2005.01457}
}
\keyword{datasets}
