\docType{package}
\name{predboot-package}
\alias{predboot}
\title{The 'predboot' package.}
\description{
A R package for computing optimism-corrected predictive accuracy measures of multivariable prediction models. It enables computations of the optimism-corrected C-statistics by bootstrap correction methods (the Harrell, 0.632, and 0.632+ methods) for logistic regression models by ML, ridge, and lasso regressions. Also, it can compute bootstrap-based optimism-corrected confidence intervals for the predictive accuracy measures proposed by Noma et al. (2020). Other functions would be added in future.
}
\author{Hisashi Noma <noma@ism.ac.jp>}
\references{
Noma, H., Shinozaki, T., Iba, K., Teramukai, S. and Furukawa, T. A. (2020). 
Confidence intervals of prediction accuracy measures for multivariable prediction models based on the bootstrap-based optimism correction methods. 
arXiv:2005.01457.
\url{https://arxiv.org/abs/2005.01457}
}
